requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
assertpy==1.1
pipenv==2022.5.2
pytest==7.1.2
selenium==4.1.5
webdriver-manager==3.5.4

